refactor(tests): move IP exam fixture to Dotsesses.Tests/TestData (TD005)
Closes TD005. Relocates the only Excel fixture the test suite actually
reads — `IP exam scores 2025.xlsx` — out of `Dotsesses/example/`
(shared with user-facing demo material) and into
`Dotsesses.Tests/TestData/`.

The duplicated `ResolveRepoFile` walk-up helpers in
`MainWindowViewModelTests`, `MainWindowViewModelBuildSeriesDataTests`,
and `StateServiceTests` (the last unused / dead code) collapse into a
single private helper on `TestFixtures`, exposed via a public
`IpExamScoresXlsx()` accessor.

`Dotsesses/example/` keeps the demo `.xlsx` outputs, the example
`.dots`, and the `.pptx` — those are user-facing example material,
not test inputs.

Docs: TECH_DEBT.md TD005
